// user_service/src/generated/user/v1/user.ts
export enum Gender {
  GENDER_UNSPECIFIED = 0,
  GENDER_MALE = 1,
  GENDER_FEMALE = 2,
  UNRECOGNIZED = -1,
}

export enum RelocationReadiness {
  RELOCATION_READINESS_UNSPECIFIED = 0,
  RELOCATION_READINESS_NOT_READY = 1,
  RELOCATION_READINESS_POSSIBLE = 2,
  RELOCATION_READINESS_READY = 3,
  UNRECOGNIZED = -1,
}

export enum ProficiencyLevel {
  PROFICIENCY_LEVEL_UNSPECIFIED = 0,
  PROFICIENCY_LEVEL_BEGINNER_A1 = 1,
  PROFICIENCY_LEVEL_ELEMENTARY_A2 = 2,
  PROFICIENCY_LEVEL_INTERMEDIATE_B1 = 3,
  PROFICIENCY_LEVEL_UPPER_INTERMEDIATE_B2 = 4,
  PROFICIENCY_LEVEL_ADVANCED_C1 = 5,
  PROFICIENCY_LEVEL_NATIVE_C2 = 6,
  UNRECOGNIZED = -1,
}

export enum EducationLevel {
  EDUCATION_LEVEL_UNSPECIFIED = 0,
  EDUCATION_LEVEL_SECONDARY = 1,
  EDUCATION_LEVEL_SPECIAL_SECONDARY = 2,
  EDUCATION_LEVEL_UNFINISHED_HIGHER = 3,
  EDUCATION_LEVEL_BACHELOR = 4,
  EDUCATION_LEVEL_MASTER = 5,
  EDUCATION_LEVEL_HIGHER = 6,
  EDUCATION_LEVEL_PHD = 7,
  UNRECOGNIZED = -1,
}
// ...
export interface UpsertUserProfileRequest {
  user_id: string;
  first_name: string;
  last_name: string;
  middle_name?: string | undefined;
  birth_date?: string | undefined;
  gender?: Gender | undefined;
  city?: string | undefined;
  photo_url?: string | undefined;
  about?: string | undefined;
  relocation_readiness?: RelocationReadiness | undefined;
  languages: LanguageInput[];
  citizenship_codes: string[];
  contacts?: UserContactInput | undefined;
}

export interface UserContactInput {
  email?: string | undefined;
  phone_number?: string | undefined;
  telegram_username?: string | undefined;
  github_username?: string | undefined;
}

export interface LanguageInput {
  code: string;
  proficiency: ProficiencyLevel;
}

export interface UserProfile {
  user_id: string;
  first_name: string;
  last_name: string;
  middle_name?: string | undefined;
  birth_date?: string | undefined;
  gender?: Gender | undefined;
  city?: string | undefined;
  photo_url?: string | undefined;
  about?: string | undefined;
  relocation_readiness?: RelocationReadiness | undefined;
  languages: LanguageProficiency[];
  citizenships: Citizenship[];
  educations: Education[];
  work_experiences: WorkExperience[];
  created_at?: Timestamp | undefined;
  updated_at?: Timestamp | undefined;
  contacts?: UserContact | undefined;
}

export interface UserContact {
  email?: string | undefined;
  phone_number?: string | undefined;
  telegram_username?: string | undefined;
  github_username?: string | undefined;
}

export interface LanguageProficiency {
  code: string;
  name: string;
  proficiency: ProficiencyLevel;
}

export interface Citizenship {
  country_code: string;
  country_name: string;
}

export interface EducationInput {
  institution_name: string;
  level: EducationLevel;
  specialization_id?: number | undefined;
  faculty?: string | undefined;
  started_at?: string | undefined;
  ended_at?: string | undefined;
}
// ...
export interface Education {
  id: string;
  institution_name: string;
  level: EducationLevel;
  specialization?: EducationSpecialization | undefined;
  faculty?: string | undefined;
  started_at?: string | undefined;
  ended_at?: string | undefined;
  created_at?: Timestamp | undefined;
  updated_at?: Timestamp | undefined;
}

export interface EducationSpecialization {
  id: number;
  name: string;
}

export interface WorkExperienceInput {
  company_name: string;
  position: string;
  description?: string | undefined;
  city?: string | undefined;
  company_url?: string | undefined;
  started_at: string;
  ended_at?: string | undefined;
}
// ...
export interface WorkExperience {
  id: string;
  company_name: string;
  position: string;
  description?: string | undefined;
  city?: string | undefined;
  company_url?: string | undefined;
  started_at: string;
  ended_at?: string | undefined;
  created_at?: Timestamp | undefined;
  updated_at?: Timestamp | undefined;
}